done.
# Arcade Game Project - A _Frogger_ Clone

## Game Instructions

The Arcade Game clone is a little game that is modelled after the well-known
game _Frogger_. You can steer the player with the arrow keys. The goal of the
game is to reach the water line.

But watch out! Don't let the bugs get you! When you collide with a bug, you have
to start again.

## Installation

Feel free to download the Arcade Game to your PC. Download the repository as a zip-file, extract it and open the index.html to run the Arcade Game.

## External Code Used in this Project

* The starter code for this project comes from the following [Udacity Arcade Game repository](https://github.com/udacity/frontend-nanodegree-arcade-game) by Richard Kalehoff. I downloaded it directly as a ZIP-file from my Udacity classroom instead of forking or cloning it. That is why you don't see a connection between the starter code and mine.
* The collision function and the switch to move the player were taken from [Antonella Dean's Arcade Game project](https://github.com/aberdean/google-scholarship-fend-projects/tree/master/classic-arcade-game-clone). I simplified the collision function and the switch to fit to the purpose of my arcade game.
* The fonts are the "Shrikhand" and "Shadows Into Light" fonts taken from [Google Fonts](https://fonts.googleapis.com/css?family=Shadows+Into+Light|Shrikhand).
* The code for the modal box was taken from [this tutorial](https://sabe.io/tutorials/how-to-create-modal-popup-box) and afterwards adapted to the needs of my Arcade Game project.